backend/requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-multipart==0.0.9


pillow==10.3.0
numpy==1.26.4
imageio==2.34.1
imageio-ffmpeg==0.4.9
av


torch==2.4.*

diffusers==0.29.*
transformers>=4.41,<5.0
accelerate>=0.30,<0.32
safetensors>=0.4.3
huggingface-hub>=0.23
